[#95] refactor: types 폴더 리팩토링
