+++ src/hooks/useGeoData.ts
@@ -1,6 +1,6 @@
 import { useEffect, useMemo, useState } from 'react';
 import { pathListFormatter } from '@/utils/region/pathListFormatter';
-import { GeoData } from '@/types/stampMap/CoordRegionCode.types';
+import { GeoData } from '@/types/stamp/coordRegionCode.type';
 import { geoListSiDoFormatter } from '@/utils/region/siDoFormatter';
 
 const useGeoData = () => {
