+++ src/serverActions/placeActions.tsx
@@ -1,17 +1,7 @@
 'use server';
 
+import { Place } from '@/types/place/place.type';
 import { createClient } from '@/utils/supabase/server';
-
-export interface Place {
-  firstimage: string | null;
-  contentid: string | null;
-  title: string | null;
-  city: string | null;
-  supabaseText: string | null;
-  citytitle: string | null;
-  citydetail: string | null;
-  isBookmarked: boolean;
-}
 
 // 데이터 호출 함수 (북마크 상태 추가)
 export const getPlaceList = async (userId: string): Promise<Place[]> => {
