+++ src/types/place/place.type.ts
@@ -1,0 +1,10 @@
+export interface Place {
+  firstimage: string | null;
+  contentid: string | null;
+  title: string | null;
+  city: string | null;
+  supabaseText: string | null;
+  citytitle: string | null;
+  citydetail: string | null;
+  isBookmarked: boolean;
+}
