--- src/types/stamp.ts
@@ -1,9 +1,0 @@
-export interface StampData {
-  id: string;
-  user_id: string;
-  region: string;
-  visited: boolean;
-  created_at: string;
-  stampimg: string;
-  location: { lat: number; lng: number };
-}
