+++ src/types/stamp/addressProps.type.ts
