+++ src/types/stamp/coordRegionCode.type.ts
