+++ src/types/stamp/stamp.type.ts
