+++ src/utils/region/pathListFormatter.ts
@@ -1,4 +1,4 @@
-import { CoordinatesType } from '@/types/stampMap/CoordRegionCode.types';
+import { CoordinatesType } from '@/types/stamp/coordRegionCode.type';
 import coordRegionCode from '../../../coordregioncode.json';
 
 // 행정구역 PathList 가져오기
